Fix small errors in merge sort java
Refactor the current interface of merge sort implementation.
Create a new access point for the sort method. The new method
requires only the array as parameter and hence solves the
problem of invalid left and right indexes (null pointer
validation is also included). The access point
is made static because no class level data are required
in the implementation. The method is exposed as public
so that it can be used from other packages.

Fixes: https://github.com/BaReinhard/Hacktoberfest-Data-Structure-and-Algorithms/issues/244