Refactor the longest common substring class
The parameters of the longest common substring methods
were modified in order to eliminate the potatial errors
caused by invalid lengths. The strings are now passed
as String objects and not char arrays. The return type
of the method was changed to return the actual substring
and not it's leght. Hence the printing is no longer needed.
The lastIndexOfY variable was removed because it was not used

Fixes: https://github.com/BaReinhard/Hacktoberfest-Data-Structure-and-Algorithms/issues/252